fix(三数/四数之和): Int溢出 → toLong() — 三四个10^9相加超Int.MAX

+++ 2607/0724/review/双指针/三数之和.kt
@@ -18,7 +18,7 @@ class ThreeSum {
             var left = i + 1
             var right = size - 1
             while (left < right) {
-                val sum = nums[i] + nums[left] + nums[right]
+                val sum = nums[i].toLong() + nums[left] + nums[right]
                 when {
                     sum == 0 -> {
                         result.add(intArrayOf(nums[i], nums[left], nums[right]).toList())
